added a claims item to the alliance menu

+++ src/org/redcastlemedia/multitallented/civs/ConfigManager.java
@@ -115,7 +115,7 @@ public class ConfigManager {
     @Getter
     List<String> allianceClaimEffects;
     @Getter
-    int powerPerAllianceClaim;
+    double powerPerAllianceClaim;
 
     public ConfigManager() {
         loadDefaults();
@@ -347,7 +347,7 @@ public class ConfigManager {
             if (allianceClaimEffects.isEmpty()) {
                 allianceClaimEffects = getDefaultAllianceEffects();
             }
-            powerPerAllianceClaim = config.getInt("power-per-alliance-claim", 1);
+            powerPerAllianceClaim = config.getDouble("power-per-alliance-claim", 1);
 
         } catch (Exception e) {
             Civs.logger.severe("Unable to read from config.yml");
